+++ level_3/fake/c/main.c
@@ -14,7 +14,7 @@
  *
  * @see exit()
  */
-int main(const unsigned long argc, const f_string_t *argv) {
+int main(const int argc, const f_string_t *argv) {
   const f_console_arguments_t arguments = { argc, argv };
   fake_data_t data = fake_data_t_initialize;
   f_status_t status = F_none;
