+++ level_3/fss_extended_write/c/main.c
@@ -1,6 +1,6 @@
 #include "fss_extended_write.h"
 
-int main(const unsigned long argc, const f_string_t *argv) {
+int main(const int argc, const f_string_t *argv) {
   const f_console_arguments_t arguments = { argc, argv };
   fss_extended_write_data_t data = fss_extended_write_data_t_initialize;
